Create specific benchmarks for useful parameterizations
